MASTER: Running.
-Version 1.0-
Wow this was a big jump! Like a few hours ago we were at V0.3 and now we're at V.10? So what changed?

Landing Page: Well it's now finished! Yes, it's pretty simple but we can add/change or do whatever with it later, but we have one running!
Contact: We have started our contact page! Sure, it doesn't have any content yet but it's alive, and we hope to add more stuff soon!